+++ R/zzz.r
@@ -13,6 +13,16 @@
 }
 
 .onAttach <- function(libname, pkgname) {
+  when_attached("data.table", {
+    if (!is_attached("dtplyr")) {
+      packageStartupMessage(rule())
+      packageStartupMessage(
+        "data.table + dplyr code now lives in dtplyr.\n",
+        "Please library(dtplyr)!"
+      )
+      packageStartupMessage(rule())
+    }
+  })
 
   setHook(packageEvent("plyr", "attach"), function(...) {
     packageStartupMessage(rule())
@@ -22,3 +32,13 @@
     packageStartupMessage(rule())
   })
 }
+
+when_attached <- function(pkg, action) {
+  if (is_attached(pkg)) {
+    action
+  } else {
+    setHook(packageEvent(pkg, "attach"), function(...) action)
+  }
+}
+
+is_attached <- function(pkg) paste0("package:", pkg) %in% search()