+++ tests/testthat/test-arrange.r
@@ -1,6 +1,6 @@
 context("Arrange")
 
-local <- c("df", "dt")
+local <- c("df")
 db <- c("sqlite", "postgres")
 
 df1 <- expand.grid(
@@ -53,21 +53,6 @@ test_that("two arranges equivalent to one", {
   tbls <- temp_load(c(local, db), df1)
   compare_tbls(tbls, ref = single, compare = equal_df,
     function(x) x %>% arrange(b) %>% arrange(a))
-})
-
-test_that("arrange results same regardless of backend", {
-  # Can't check db because types are not currently preserved
-  tbls <- temp_load(local, df2)
-
-  compare_tbls(tbls, function(x) x %>% arrange(a, id), compare = equal_df)
-  compare_tbls(tbls, function(x) x %>% arrange(b, id), compare = equal_df)
-  compare_tbls(tbls, function(x) x %>% arrange(c, id), compare = equal_df)
-  compare_tbls(tbls, function(x) x %>% arrange(d, id), compare = equal_df)
-
-  compare_tbls(tbls, function(x) x %>% arrange(desc(a), id), compare = equal_df)
-  compare_tbls(tbls, function(x) x %>% arrange(desc(b), id), compare = equal_df)
-  compare_tbls(tbls, function(x) x %>% arrange(desc(c), id), compare = equal_df)
-  compare_tbls(tbls, function(x) x %>% arrange(desc(d), id), compare = equal_df)
 })
 
 test_that("arrange handles list columns (#282)", {