+++ tests/testthat/test-do.R
@@ -8,7 +8,7 @@ df <- data.frame(
   y = 6:1
 )
 
-srcs <- temp_srcs(c("df", "dt", "sqlite"))
+srcs <- temp_srcs(c("df", "sqlite"))
 tbls <- temp_load(srcs, df)
 grp <- lapply(tbls, function(x) x %>% group_by(g))
 
@@ -139,56 +139,6 @@ test_that("empty data frames give consistent outputs", {
     expect_equal(c(x = "dbl", g = "chr", y = "int"))
   emt %>% do(y = ncol(.)) %>% type_sum() %>%
     expect_equal(c(g = "chr", y = "list"))
-})
-
-# Data tables  -----------------------------------------------------------------
-
-test_that("ungrouped data table with unnamed argument returns data frame", {
-  out <- tbl_dt(mtcars) %>% do(head(.))
-  expect_is(out, c("tbl_dt", "data.table"))
-  expect_equal(dim(out), c(6, 11))
-})
-
-test_that("ungrouped data frame with named argument returns list data frame", {
-  out <- tbl_dt(mtcars) %>% do(x = 1, y = 2:10)
-  expect_is(out, c("tbl_dt", "data.table"))
-  expect_equal(out$x, list(1))
-  expect_equal(out$y, list(2:10))
-})
-
-test_that("named argument become list columns", {
-  out <- grp$dt %>% do(nrow = nrow(.), ncol = ncol(.))
-  expect_equal(out$nrow, list(1, 2, 3))
-  # doesn't including grouping columns
-  expect_equal(out$ncol, list(3, 3, 3))
-})
-
-test_that("unnamed results bound together by row", {
-  first <- grp$dt %>% do(head(., 1))
-
-  expect_equal(nrow(first), 3)
-  expect_equal(first$g, 1:3)
-  expect_equal(first$x, c(1, 2, 4))
-})
-
-test_that("grouped_dt do evaluates args in correct env", {
-  a <- 10
-  f <- function(a) {
-    grp$dt %>% do(a = a)
-  }
-
-  expect_warning(x <- f(20)$a)
-  expect_equal(x, list(20, 20, 20))
-})
-
-test_that("grouped_dt passes all columns", {
-  out <- mtcars %>%
-    tbl_dt() %>%
-    select(mpg, cyl) %>%
-    group_by(cyl) %>%
-    do(n = names(.))
-
-  expect_equal(out$n[[1]], c("mpg", "cyl"))
 })
 
 # SQLite -----------------------------------------------------------------------